Join scores with unequal number of staves
The first score defines the order of parts and staves, so it must
contain all parts and staves that appear in any of the later scores.
However, later scores do not need to contain all parts or staves that
are present in the first score. If a staff is missing from a later
score then it will be assumed tacet and padded with rests in the
joined score.