Fix instrument names in template scores
MuseScore has "D Trumpet" and "Horn in D". Notice how the key can come
before or after the instrument name. This commit chooses the latter
style. It also numbers instruments if there is more than one of the
same type in the score.

If a score contains instruments like:

- D Trumpet
- D Trumpet
- Horn in B♭ basso
- Horn in B♭ basso

Then the names will be adjusted like:

- Trumpet 1 in D
- Trumpet 2 in D
- Horn 1 in B♭ basso
- Horn 2 in B♭ basso